feat: custom relationship types (#20)
When serializing data with relationships, the relationship type is
currently set to whatever the key on the data object is. This isn't
always desired. We could have two relationship properties that are of
the same type, but they cannot share a key in typescript.

Change the serialize relationships property to accept a string array of
property names _or_ a mapping from property name to the related
resource's type.